Renamed Main class to Forwarder
